milestone3/requirements.txt
streamlit
nltk
spacy
PyPDF2
python-docx